+++ tool/tools.go
@@ -72,15 +72,17 @@ func GetAddressFromPubKey(pubKey string) (address string, err error) {
 		log.Println(err)
 		return "", errors.New("invalid pubKey")
 	}
-	// test TestNet3Params
+
 	// main MainNetParams
 	var net *chaincfg.Params
 	if strings.Contains(config.ChainNode_Type, "main") {
 		net = &chaincfg.MainNetParams
 	}
+	// test TestNet3Params
 	if strings.Contains(config.ChainNode_Type, "test") {
 		net = &chaincfg.TestNet3Params
 	}
+	// reg RegressionNetParams
 	if strings.Contains(config.ChainNode_Type, "reg") {
 		net = &chaincfg.RegressionNetParams
 	}
